update to the latest gonum
The old version of gonum that this was built on did not have a go.mod
file and so it was impossible to lock the version of the dependency. The
newer version seemed to have changed a few method signatures and this
change addresses that as well. The biggest change was that when getting
the eigen vectors, we now get a complex matrix instead of a matrix with
just real numbers. After some testing, we believe that the imaginary
parts are always zero, so this code converts the complex matrices to
real matrices in order to do the same math we did before.